Revert "Wrap iterator-advancement invocation (used to retrieve measurement-units) within a try...except handler, and instead of providing a default should the iterator be exhausted, handle the resulting StopIteration exception instead"
This reverts commit 9beec3cb8fba1a8596e552027fc57b4ae4cfbba4.

Rationale: although it _shouldn't_ be possible for the 'return' statement within the 'except' block to be reached (because all designator values in the token iterators are followed by a unit measurement name), it makes the number of future changes for that to _potentially_ occur smaller.  For example, someone could claim that for some reason it'd make sense to remove the final measurement name from each token iterator.  Slightly implausible -- and maybe not the best example; someone clever could probably come up with a better one -- but if it's possible to increase the distance between valid-code and problematic-code (without sacrificing much readability and without much additional code/runtime/other overheads), then that seems worth doing.